重构库文件下载
