+++ src/main/resources/libraries.info
@@ -1,0 +1,57 @@
+source|http://sv.catserver.moe:8001/dl/|null
+source|http://sv2.catserver.moe:8001/dl/|null
+source|http://cdn.catserver.moe/dl/|null
+lib|minecraft_server.1.12.2.jar|71728ed3fbd0acd1394bf3ade2649a5c
+lib|libraries/snakeyaml-1.19.jar|95472b5a0ded8761545342a087e82117
+lib|libraries/SpecialSource-1.7.4.jar|eded0fa5002ee8e12b06552ebee76c99
+lib|libraries/sqlite-jdbc-3.21.0.1.jar|ae5064ef5ef5005e6c7e5bb95fd8e9ba
+lib|libraries/trove4j-3.0.3.jar|8fc4d4e0129244f9fd39650c5f30feb2
+lib|libraries/vecmath-1.5.2.jar|e5d2b7f46c4800a32f62ce75676a5710
+lib|libraries/akka-actor_2.11-2.3.3.jar|541440ca0819ebada47d6d1a8b3ee9e1
+lib|libraries/asm-debug-all-5.2.jar|fe5f20404ccdee9769ef05dc4b47ba98
+lib|libraries/authlib-1.5.25.jar|321723be87583ff5f427a37d74081370
+lib|libraries/bungeecord-chat-1.12-SNAPSHOT.jar|3227092125953bd4b82e80958e24695f
+lib|libraries/codecjorbis-20101023.jar|0d622e2ac4368b5a33d540a9e4819e0c
+lib|libraries/commons-codec-1.10.jar|353cf6a2bdba09595ccfa073b78c7fcb
+lib|libraries/commons-compress-1.8.1.jar|d862e30ff6b5d78264677dcd6507abb8
+lib|libraries/commons-io-2.5.jar|e2d74794fba570ec2115fb9d5b05dc9b
+lib|libraries/commons-lang-2.6.jar|4d5c1693079575b362edf41500630bbd
+lib|libraries/commons-lang3-3.5.jar|780b5a8b72eebe6d0dbff1c11b5658fa
+lib|libraries/commons-logging-1.1.3.jar|92eb5aabc1b47287de53d45c086a435c
+lib|libraries/config-1.3.1.jar|1d0abce257a747570923243f3105ffca
+lib|libraries/fastutil-7.1.0.jar|35fc1f3aaab7a782873be02319a53828
+lib|libraries/gson-2.8.0.jar|a42f1f5bfa4e6f123ddcab3de7e0ff81
+lib|libraries/guava-21.0.jar|ddc91fd850fa6177c91aab5d4e4d1fa6
+lib|libraries/hamcrest-core-1.1.jar|b66d0c48e1f1dc54d4227db52512c15b
+lib|libraries/httpclient-4.3.3.jar|88cc3123fce88d61b7c2cdbfc33542c5
+lib|libraries/httpcore-4.3.2.jar|ee3d34dce4a30c7d3002cadf8c9172c1
+lib|libraries/icu4j-core-mojang-51.2.jar|aec124acf7b3c1c6ed41a6270a4452b8
+lib|libraries/jansi-1.11.jar|e8bd19df14afe8a0f4e2a44d57c0cd8b
+lib|libraries/java-objc-bridge-1.0.0.jar|2003e6c279b09f657509aa038329aa7b
+lib|libraries/jline-2.13.jar|f251ba666cccb260ff7215b2cbeee8d4
+lib|libraries/jna-4.4.0.jar|34d3537524a6c8c134e840e7be601569
+lib|libraries/jopt-simple-5.0.3.jar|0a5ec84e23df9d7cfb4063bc55f2744c
+lib|libraries/json-simple-1.1.1.jar|5cc2c478d73e8454b4c369cee66c5bc7
+lib|libraries/jsr305-3.0.1.jar|c6532beb3f7cc54a8d73d25d5602b9e4
+lib|libraries/junit-4.10.jar|68380001b88006ebe49be50cef5bb23a
+lib|libraries/jutils-1.0.0.jar|f60976b19661c849c5c87433045a9885
+lib|libraries/launchwrapper-1.12.jar|934b2d91c7c5be4a49577c9e6b40e8da
+lib|libraries/log4j-api-2.8.1.jar|a2ad9b058b4b03d43f3cc301701654e4
+lib|libraries/log4j-core-2.8.1.jar|547bb3ed2deb856d0e3bbd77c27b9625
+lib|libraries/lzma-0.0.1.jar|a3e3c3186e41c4a1a3027ba2bb23cdc6
+lib|libraries/maven-artifact-3.5.3.jar|7741ebf29690ee7d9dde9cf4376347fc
+lib|libraries/mysql-connector-java-5.1.45.jar|7e5fe17c66a6c227c4747ab383a492d0
+lib|libraries/netty-all-4.1.9.Final.jar|2f4a891dffdb1259e4b10f1f090224e2
+lib|libraries/oshi-core-1.1.jar|4f992d3ac0aa70a8647460494c95e261
+lib|libraries/platform-3.4.0.jar|c5057d6dc3a531708e4a91249fadda3f
+lib|libraries/plexus-utils-3.1.0.jar|bfec331a62402081dd4143e3a8d193e4
+lib|libraries/scala-actors-2.11.0.jar|0c6b3b08bb49fbe168089bd4438937bc
+lib|libraries/scala-actors-migration_2.11-1.1.0.jar|f5e79398daa1806f8b17311a3c782723
+lib|libraries/scala-compiler-2.11.1.jar|06030143bf86ca896fb6ccfd679b5760
+lib|libraries/scala-continuations-library_2.11-1.0.2.jar|004d7007abbcee858d3ca2c3ccbcbaab
+lib|libraries/scala-continuations-plugin_2.11.1-1.0.2.jar|359c4a6743a082c689039482eed78670
+lib|libraries/scala-library-2.11.1.jar|1d88f665219e6006c5dd82d71c525c0f
+lib|libraries/scala-parser-combinators_2.11-1.0.1.jar|4e694499c965af4a02599c99d4f0b196
+lib|libraries/scala-reflect-2.11.1.jar|7878fac044e4e4b576bb35a77ccc34fc
+lib|libraries/scala-swing_2.11-1.0.1.jar|1009d69e4948045383f2a7a334348af5
+lib|libraries/scala-xml_2.11-1.0.2.jar|c2d7e66495afe14545c31b21e99879ef
